fix(blobLength): use uint32 instead of uint
We've been using uint/int in a bunch of data types which are fundamental to our protocol, and influence hashing and possibly other things. We shouldn't.

bloblength is once instance where we use uint32 onchain, uint offchain (and actually uint64 in the encoder proto!).

This PR changes the main offchain bloblength type from uint to uint32, and propagates the changes everywhere.

I decided not to change all the v1 chunk encoding/assignment logic which still uses uint, and so just convert uint32->uint when calling those functions. This is fine because golang uint type is always at least uint32.

+++ core/v2/types.go
@@ -143,7 +143,7 @@ func BlobHeaderFromProtobuf(proto *commonpb.BlobHeader) (*BlobHeader, error) {
 			Commitment:       commitment,
 			LengthCommitment: lengthCommitment,
 			LengthProof:      lengthProof,
-			Length:           uint(proto.GetCommitment().GetLength()),
+			Length:           proto.GetCommitment().GetLength(),
 		},
 		QuorumNumbers:   quorumNumbers,
 		PaymentMetadata: *paymentMetadata,
@@ -169,8 +169,8 @@ func (b *BlobHeader) ToProtobuf() (*commonpb.BlobHeader, error) {
 	}, nil
 }
 
-func GetEncodingParams(blobLength uint, blobParams *core.BlobVersionParameters) (encoding.EncodingParams, error) {
-	length, err := blobParams.GetChunkLength(uint32(blobLength))
+func GetEncodingParams(blobLength uint32, blobParams *core.BlobVersionParameters) (encoding.EncodingParams, error) {
+	length, err := blobParams.GetChunkLength(blobLength)
 	if err != nil {
 		return encoding.EncodingParams{}, err
 	}
